Move extension methods to instance methods
