+++ src/ModelContextProtocol.Core/Client/McpClientTool.cs
@@ -19,8 +19,8 @@ namespace ModelContextProtocol.Client;
 /// <see cref="WithName"/> and <see cref="WithDescription"/> without changing the underlying tool functionality.
 /// </para>
 /// <para>
-/// Typically, you would get instances of this class by calling the <see cref="McpClientExtensions.ListToolsAsync"/>
-/// or <see cref="McpClientExtensions.EnumerateToolsAsync"/> extension methods on an <see cref="McpClient"/> instance.
+/// Typically, you would get instances of this class by calling the <see cref="McpClient.ListToolsAsync"/>
+/// or <see cref="McpClient.EnumerateToolsAsync"/> extension methods on an <see cref="McpClient"/> instance.
 /// </para>
 /// </remarks>
 public sealed class McpClientTool : AIFunction
