+++ src/ModelContextProtocol.Core/Protocol/ClientCapabilities.cs
@@ -44,7 +44,7 @@ public sealed class ClientCapabilities
     /// server requests for listing root URIs. Root URIs serve as entry points for resource navigation in the protocol.
     /// </para>
     /// <para>
-    /// The server can use <see cref="McpServerExtensions.RequestRootsAsync"/> to request the list of
+    /// The server can use <see cref="McpServer.RequestRootsAsync"/> to request the list of
     /// available roots from the client, which will trigger the client's <see cref="RootsCapability.RootsHandler"/>.
     /// </para>
     /// </remarks>
