+++ src/ModelContextProtocol.Core/Protocol/ITransport.cs
@@ -63,7 +63,7 @@ public interface ITransport : IAsyncDisposable
     /// <para>
     /// This is a core method used by higher-level abstractions in the MCP protocol implementation.
     /// Most client code should use the higher-level methods provided by <see cref="McpSession"/>,
-    /// <see cref="McpSessionExtensions"/>, <see cref="McpClientExtensions"/>, or <see cref="McpServerExtensions"/>,
+    /// <see cref="McpClient"/>, or <see cref="McpServer"/>,
     /// rather than accessing this method directly.
     /// </para>
     /// </remarks>
